Improve plotter output, introduce randomness (#135)
This improves the `sample` function central to Mafs' plotter.

### Bug fix: duplicate points, incorrect subdivision

First, it corrects a bug that was causing duplicate points to be
evaluated. For example, when sampling the domain [-16, 16] with sampling
depth 4, these points were being generated:

```
[-16,-15,-13,-12,-12,-11,-9,-8,-8,-7,-5,-4,-4,-3,-1,0,0,1,3,4,4,5,7,8,8,9,11,12,12,13,15,16]
```

Of course, duplicates make no sense here, and we would actually expect a
uniform distribution between -16 and 16. Also, we would expect depth _5_
to produce 32 segments (33 points), not depth 4. So that was fixed by,
thankfully, removing code.

### Randomness

Additionally, for quite some time, I have been wanting to introduce some
stable randomness to the sampling function, in a way that prevents the
function from changing with every render. I finally decided to research
that and introduce a pseudo-random hash function when evaluating the
midpoint of the domain. This also required the `midpoint` function in
`sample` to become a `lerp`, since we can no longer simply measure
against the midpoint to compute errors.