Extract useMovementInteraction and MovablePointSVG from MovablePoint
Mafs comes with built-in support for movable points via the MovablePoint
component. These points can be dragged with the mouse or moved with the
arrow keys. However, we'd like to make other kinds of elements (line
segments, polygons) movable too, in a way that feels consistent with the
movement of points.

To achieve this consistency, we'd like to reuse the movement logic from
`MovablePoint`. To that end, we export a `useMovementInteraction` hook
from Mafs.

We also split out a presentational component that just renders the
MovablePoint SVG. We think we might want to create our own keyboard
movement logic in the future, with different meanings for the ctrl and
alt modifiers. Having a purely presentational component for movable
points will facilitate that.

Issue: https://khanacademy.atlassian.net/browse/LEMS-1672

Test plan:

```
yarn start
open http://localhost:3000
```

Navigate to the Movable Point page and play with the examples.

There are also Playwright e2e tests that cover MovablePoint.
To run them:

```bash
npx playwright install # one-time setup to globally install browsers
pnpm test
```