Extract hook and display component from MovablePoint (#144)
We at @khan would like to add movable circles, polygons, and line
segments
to Mafs graphs. For the sake of keeping the UX consistent, we'd like the
movement interaction for these objects to reuse the hook from
MovablePoint,
but that hook currently isn't exported.

We're also thinking that we may want to customize e.g. the size of the
movable points.

To that end, this PR splits MovablePoint into a hook and a
presentational
component (both exported), so that these aspects can be reused and
changed
independently.

@stevenpetryk I'd welcome any feedback you have. I'm especially looking
for
feedback on naming and code organization.

---------

Co-authored-by: Steven Petryk <petryk.steven@gmail.com>