Enable Setting of Domain for X & Y Parametric Plots (#166)
There are times when a user may want to restrict a plot via a set of
domain parameters. The current `<Parametric>` component already limits
the plot to whatever fits within the bounds of the graph. By adding an
optional `domain` property to the `OfX` and `OfY` versions of the
`<Parametric>` component, we can use those values when determining the
domain (formerly `t`) of the plot.

As part of this expansion of the domain property, we are deprecating the
`t` property in the `<Parametric>` component and replacing it with
`domain`. The functionality is the same, but the property name better
conveys its purpose.

---------

Co-authored-by: Mark Fitzgerald <markfitzgerald@markfitgeraldm3.home>
Co-authored-by: Steven Petryk <petryk.steven@gmail.com>